Made similar screens to reuseable component
