+++ src/constants/Forum.js
@@ -1,4 +1,4 @@
-export const manageForumDetailsType = {
+export const renderType = {
   CREATE: 0,
   EDIT: 1
 };
