+++ src/screens/Forum/CreateForumScreen.js
@@ -1,10 +1,8 @@
 import RenderManageForumDetails from "../../components/Forum/RenderManageForumDetails";
-import { manageForumDetailsType } from "../../constants/Forum";
+import { renderType } from "../../constants/Forum";
 
 const CreateForumScreen = () => {
-  return (
-    <RenderManageForumDetails manageType={manageForumDetailsType.CREATE} />
-  );
+  return <RenderManageForumDetails manageType={renderType.CREATE} />;
 };
 
 export default CreateForumScreen;
